feature/prepare config to testing in jest and supertest
