+++ package.json
@@ -14,7 +14,9 @@
       "@types/node": "^16.11.10",
       "@types/swagger-jsdoc": "^6.0.2",
       "@types/swagger-ui-express": "^4.1.5",
+      "jest": "^29.7.0",
       "nodemon": "^3.0.1",
+      "supertest": "^6.3.3",
       "ts-node": "10.7.0",
       "typescript": "4.5.2"
    },
